refactor: split ssh host management api
